tried adding an integer display and memory size detector function

+++ drivers/mem.h
@@ -1,1 +1,2 @@
-long getMemSize();
+
+long long getMemSize();
